+++ apps/web/src/App.tsx
@@ -16,6 +16,7 @@ import { HouseholdIncomePage } from './pages/HouseholdIncomePage'
 import { BudgetYearsPage } from './pages/BudgetYearsPage'
 import { ComparePage } from './pages/ComparePage'
 import { SavingsPage } from './pages/SavingsPage'
+import { HistoryPage } from './pages/HistoryPage'
 
 const queryClient = new QueryClient()
 
@@ -119,6 +120,14 @@ function App() {
               element={
                 <ProtectedRoute>
                   <ComparePage />
+                </ProtectedRoute>
+              }
+            />
+            <Route
+              path="/households/:id/history"
+              element={
+                <ProtectedRoute>
+                  <HistoryPage />
                 </ProtectedRoute>
               }
             />